pyproject.toml
[tool.poetry]
name = "properties-prices-register"
version = "0.1.0"
description = ""
authors = ["Lucio Tudisco <lucio.tudisco@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.1"
django-celery-results = "^2.5.1"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
redis = "^5.0.1"
requests = "^2.31.0"
types-requests = "^2.31.0.20231231"
djangorestframework = "^3.14.0"
markdown = "^3.5.1"
django-filter = "^23.5"
googlemaps = "^4.10.0"

[tool.poetry.group.dev.dependencies]
pre-commit = "^3.6.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.mypy.overrides]]
module = "*.migrations.*"
ignore_errors = true
